+++ cloud-consume/src/main/java/com/ybin/consumer/controller/Consumer.java
@@ -12,12 +12,11 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.client.RestTemplate;
 
 /**
  * @Describe:
- * @Author: bing.yue001
+ * @Author: bing.yue
  */
 @RequestMapping(value = "/consumer")
 @Controller
