--- cloud-learning-provider/cloub-learning-service-backup/src/main/java/com/ybin/provider/service/ApplicationProviderBackUpService.java
@@ -1,24 +1,0 @@
-package com.ybin.provider.service;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
-import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-import org.springframework.context.annotation.EnableAspectJAutoProxy;
-
-/**
- * @author ybin
- */
-@EnableAspectJAutoProxy
-@EnableEurekaClient
-@EnableFeignClients
-@SpringBootApplication(exclude = DataSourceAutoConfiguration.class)
-public class ApplicationProviderBackUpService {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ApplicationProviderBackUpService.class, args);
-	}
-
-}
-
