--- cloud-learning-provider/cloub-learning-service-backup/src/main/resources/bootstrap.properties
@@ -1,5 +1,0 @@
-spring.application.name=cloud-service
-server.port=1013
-
-eureka.client.service-url.defaultZone=http://127.0.0.1:1010/eureka/
-eureka.instance.instance-id=${spring.cloud.client.ip-address}:${server.port}:${spring.application.name}
