+++ cloud-provider-service/cloud-provider-api/src/main/java/com/ybin/provider/api/consumer/domain/ConsumerBo.java
@@ -1,0 +1,24 @@
+package com.ybin.provider.api.consumer.domain;
+
+import lombok.Data;
+
+/**
+ * @Describe:
+ * @Author: bing.yue
+ */
+@Data
+public class ConsumerBo {
+
+    /**
+     * id
+     */
+    private String id;
+    /**
+     * 服务名
+     */
+    private String serverName;
+    /**
+     * 端口
+     */
+    private String port;
+}
