+++ cloud-provider-service/cloud-provider/src/main/java/com/ybin/provider/service/ApplicationProviderService.java
