+++ cloud-provider-service/cloud-provider/src/main/resources/bootstrap.properties
