+++ cloud-provider-service/pom.xml
@@ -9,7 +9,13 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>cloud-consume</artifactId>
+    <artifactId>cloud-provider-service</artifactId>
+    <packaging>pom</packaging>
+    <modules>
+        <module>cloud-provider-api</module>
+        <module>cloud-provider</module>
+        <module>cloud-provider-backup</module>
+    </modules>
 
     <dependencies>
         <dependency>
@@ -40,28 +46,7 @@
             <artifactId>spring-boot-starter-test</artifactId>
             <scope>test</scope>
         </dependency>
-
-        <dependency>
-            <groupId>org.springframework.cloud</groupId>
-            <artifactId>spring-cloud-starter-config</artifactId>
-            <version>2.0.3.RELEASE</version>
-        </dependency>
-
-        <!--<dependency>-->
-            <!--<groupId>org.springframework.boot</groupId>-->
-            <!--<artifactId>spring-boot-starter-actuator</artifactId>-->
-        <!--</dependency>-->
-        <dependency>
-            <groupId>com.ybin</groupId>
-            <artifactId>cloud-provider-api</artifactId>
-            <version>1.0-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>com.alibaba</groupId>
-            <artifactId>fastjson</artifactId>
-            <version>1.2.28</version>
-        </dependency>
-
     </dependencies>
+
 
 </project>
