+++ pom.xml
@@ -9,7 +9,7 @@
     <packaging>pom</packaging>
     <version>1.0-SNAPSHOT</version>
     <modules>
-        <module>cloud-learning-provider</module>
+        <module>cloud-provider-service</module>
         <module>cloud-eureka</module>
         <module>cloud-consume</module>
         <module>cloud-zuul</module>
